Add variable resources to httpserver
Each Route can now have a single "variable" route that will match any
request entered that doesn't match any of the other children. The call
method for routes have been changed so that the varibale data entered
into the url can be accessed by the handlers. This now means that the
router will call the base route rather than finding the route and
calling it directly.